backend/Services/co_scientist_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
PyYAML
networkx
sentence-transformers
qdrant-client
faiss-cpu
spacy
scispacy
biopython
pdfplumber
PyMuPDF
prometheus-client
opentelemetry-api
opentelemetry-sdk
httpx
pytest
streamlit
plotly
